In t2grid minc(), make sure volume_fractions array is float (in case of integer input)

+++ t2grids.py
@@ -888,7 +888,7 @@ class t2grid(object):
             from scipy.misc import derivative
             from numbers import Number
 
-            volume_fractions = np.array(volume_fractions)
+            volume_fractions = np.array(volume_fractions, dtype = np.float)
 
             if isinstance(spacing, Number): spacing = [spacing]
             missing = num_fracture_planes - len(spacing)
